feat: add new validator to only check the usage of deprecate apis (#156)
* feat: add new validator to only check the usage of deprecate apis

* applying review sugesstions

Upstream-repository: api
Upstream-commit: e2541569a535e6421048bfaee00b781bb3dc092f
